Day_07_Sampling_From_DDPM/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ddpm-sampling-day7"
version = "0.1.0"
description = "Day 7: DDPM Sampling Implementation"
requires-python = ">=3.8"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "numpy>=1.21.0",
    "matplotlib>=3.5.0",
    "Pillow>=9.0.0",
    "PyYAML>=6.0",
    "tqdm>=4.60.0",
    "imageio>=2.25.0",
    "lpips>=0.1.4",
    "scipy>=1.9.0",
    "pytest>=7.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
